Move normalizeMdx to helpers/mdx
