+++ src/helpers/mdx.js
@@ -1,3 +1,5 @@
+import { DOCS_CONTENT_URL } from "helpers/documentation";
+
 export const extractTextFromAst = (ast) =>
   ast.children
     .flatMap(handleNode)
@@ -41,4 +43,58 @@ export const getDescriptionFromAst = (ast) => {
     }
     return newDescription;
   }, "");
+};
+
+/**
+ *
+ * @param {object} node a Gatsby MDX node.
+ * @param {string} node.id MDX id
+ * @param {object} node.frontmatter frontmatter
+ * @param {string} node.frontmatter.order order
+ * @param {string} node.frontmatter.title title
+ * @param {string} node.body body
+ * @param {object} node.parent parent
+ * @param {string} node.parent.relativeDirectory relativeDirectory
+ * @param {string} [node.parent.relativePath] relativePath
+ * @param {object} [node.parent.fields] fields
+ * @param {object} [node.parent.fields.metadata.data.order] order
+ * @param {object} [node.parent.fields.metadata.data.title] title
+ *
+ * @returns {object} An object with the below keys. Any missing fields on the
+ * input will produce `undefined`. on the output.
+ * id
+ * order
+ * title
+ * githubLink
+ * body
+ * directory
+ * currentDirectory
+ * folder: {
+ *   order
+ *   title
+ * }
+ * ...node
+ */
+export const normalizeMdx = (node) => {
+  const { id, frontmatter = {}, body, parent = {} } = node;
+  const metadata = parent.fields?.metadata.data || {};
+  const parentRelativeDirSplit = parent.relativeDirectory?.split("/");
+  const mdxLink = parent.relativePath && DOCS_CONTENT_URL + parent.relativePath;
+
+  return {
+    ...node,
+    id,
+    order: frontmatter.order,
+    title: frontmatter.title,
+    githubLink: mdxLink,
+    body,
+    // paths always start with `docs` or `api` which isn't useful. Strip it.
+    directory: parentRelativeDirSplit?.slice(1).join("/"),
+    currentDirectory:
+      parentRelativeDirSplit?.[parentRelativeDirSplit.length - 1],
+    folder: {
+      order: metadata.order,
+      title: metadata.title,
+    },
+  };
 };
